Rename packages to be more consistent; add tileentity-package
